- Changed access policy of TableComposite logic (to make it implementation specific) - Transaction modification events support (changed tables on COMMIT command) - Binary serialization of Types (to write table headers) - Binary serialization of Expressions (only FunctionCall supported now) - Fixes to the table identity (TableId) determination

+++ src/deveeldb/Deveel.Data.Index/IMappedBlock.cs
@@ -20,9 +20,9 @@ using Deveel.Data.Store;
 
 namespace Deveel.Data.Index {
 	interface IMappedBlock : IIndexBlock<int> {
-		int FirstEntry { get; }
+		long FirstEntry { get; }
 
-		int LastEntry { get; }
+		long LastEntry { get; }
 
 		long BlockPointer { get; }
