+++ src/deveeldb/Deveel.Data.Transactions/TableCommitInfo.cs
@@ -20,7 +20,7 @@ using System.Linq;
 
 namespace Deveel.Data.Transactions {
 	public sealed class TableCommitInfo {
-		public TableCommitInfo(int commitId, ObjectName tableName, IEnumerable<int> addedRows, IEnumerable<int> removedRows) {
+		public TableCommitInfo(long commitId, ObjectName tableName, IEnumerable<int> addedRows, IEnumerable<int> removedRows) {
 			if (tableName == null)
 				throw new ArgumentNullException("tableName");
 
@@ -33,7 +33,7 @@ namespace Deveel.Data.Transactions {
 				RemovedRows = removedRows.ToList();
 		}
 
-		public int CommitId { get; private set; }
+		public long CommitId { get; private set; }
 
 		public ObjectName TableName { get; private set; }
